non-upstream: rewrite feat_flag to preven the use of BIT MACRO overflow
while feat_flag is getting more and more,
the use of the MACRO "BIT" will hit its upper bound depends on platform.
to prevent this in the future.
changing the writing method of feat_flag.

the driver module parameter "feature_disable" is also rewrote
accordingly.
the new usage of this parameter is as below,

insmod brcmfmac.ko debug=0x100006 feature_disable=0xff,0x03

This would set the first byte to 0xff (all features in the first byte disabled)
and the second byte to 0x03 (first two features in the second byte disabled).

Signed-off-by: Carter Chen <carter.chen@infineon.com>