brcmfmac: porting auth_status v2 for WPA3 roam fails
[Analysis]:
8950/H2 STA WPA3 roam fails because
when wpa3 external auth request event is sent.

the supplicant might not have the scan result of the target AP
as it is roam scan is triggered independent of supplicant.
If the supplicant doesn't find the Target AP in its scan results,
it uses Hunting and Pecking by default for SAE.

even when the WPA3 Target AP is supporting H2E.
it is leading to auth rejection from the Target AP.

[Fix]:
Update bss information of the Target AP with
which we are starting external auth to the cfg80211 layer.
If the supplicant doesn't find the Target AP in its scan results
it queries the cfg80211 layer for scan results.
Update new structure wl_auth_req_status_info_v2_t
for the auth request status event which
will carry the bss information of the target AP.

Signed-off-by: Carter Chen <carter.chen@infineon.com>