non-upstream: assigned an non-zero packetid to prevent fw misupdate
the root cause of SWLINUX-5091 is found.

the symptom is the SAP of H1 extsae sends out eapol msg1 that
has around 1~2 seconds delay after the DUT SAP replied the assoc resp.

it dues to the firmware reports the mgmt_txstatus event of auth packet
to the host driver that is with a unmatch packet id.
which causes the driver doesn't call complete function to finish the waiting of mgmt_tx.

why does the firmware reports a wrong packet ID?
it dues to there is an union data structure in the pkttag.
if the firmware enabled WLPKTDLYSTAT compile option.
the packetid and enqtime are sharing the same memory space.

so, the flow is like this.
The driver fills mgmt_tx for auth resp packet.
fill the packet id is 0 from the host.

the firmware takes a pkt and fill the packet id is 0 as well.
the packet passes through txq_hw_hdr_fixup,

it checks the exqtime(sharing the same memory space of packetid), then update it.
then the issue happen.

sync the host driver filled a non-zero value to packetid from DHD.

Fixes SWLINUX-5091

Signed-off-by: Carter.Chen <carter.chen@infineon.com>