brcmfmac: Low WMM TX Throughput fix for BK access category
The continuous addition of new features and fixes has resulted in a
progressive increase in firmware size. This expansion directly impacts
the available memory resources, leading to a reduction in the overall
packet pool size.

In the current implementation, the Background traffic category being
lower priority receives a smaller share of resources compared to
higher-priority access categories.

Furthermore, the existing driver design has a limitation that prohibits
borrowing credits for the BK traffic category. This design constraint
results in significantly reduced throughput for BK traffic.

To address this issue and improve the throughput for BK traffic,
the driver has been updated to allow borrowing of credits for BK access
category from higher-priority access categories.

Signed-off-by: avishad verma <avishad.verma@infineon.com>

Co-authored-by: avishad verma <avishad.verma@infineon.com>