brcmfmac: handle SDIO readl command error.
In the current implementation, the sdio_readl command can fail and
return 0xFFFFFFFF. However, in several instances within the existing code,
there is no explicit validation to determine whether the sdio_readl
operation was successful before using its return value.
This oversight leads to scenarios where the returned 0xFFFFFFFF value is
interpreted as valid data, causing driver to execute incorrect operations.
One such scenario occurs when the driver detects a firmware (FW) crash.
In reality, the observed behavior is due to the failure of the sdio_readl
command rather than an actual FW crash.
The use of 0xFFFFFFFF as a valid return value misleads the driver logic,
resulting in erroneous conclusions.
To mitigate this, the return value of sdio_readl is modilied to 0 in
failure scenarios.

Signed-off-by: avishad verma <avishad.verma@infineon.com>